added a test to reproduce the issue #222 (it fails as reported when using the StdHttpClient.Builder) + provided an ApacheHC configuration that make this test pass.
The configuration is provided as a Spring Context for conveniance because this is the configuration I am using for a long time in my application code (it can easily be changed to Java code).

It is a work in progress, as a first step to save this work, it depends on classes that have long been in my repository only, so it needs a cleanup before I can make a PR upstream.